+++ apps/users/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin as BaseUserAdmin
-from .models import User, UserProfile, UserPrivateData, Badge
+from apps.users.models import User, UserProfile, UserPrivateData, Badge
 
 # 1. Konfigurasi Inline untuk Profile
 class UserProfileInline(admin.StackedInline):