+++ apps/users/views/__init__.py
@@ -1,0 +1,2 @@
+from .authentications import RegisterView, LoginView, LogoutView
+from .profiles import MyProfileView